% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainCode.R
\name{plot.olbm}
\alias{plot.olbm}
\title{Plot OLBM}
\usage{
\method{plot}{olbm}(x, type = "hist", ...)
}
\arguments{
\item{x}{The "olbm" object output of the function olbm.}

\item{type}{A string specifying the type of plot to be produced. The currently supported values are "hist" and "incidence".}

\item{...}{Additional parameters to pass to sub-functions.}
}
\description{
It plots the re-organized incidence matrix and/or the estimated Gussian densities.
}
\examples{
data(olbm_dat)
res <- olbm(olbm_dat$Y, Q=3, L=2)   
plot(res, "hist")
plot(res, "incidence")
}
